QTextOption: fix handling Qt::AlignBaseline
QTextOption stores the Qt::Alignment flag in a eight bit bitfield. When
Qt::AlignBaseline was added the bitfield was not extended so the newly
used ninth bit is not stored in QTextOption.
Fix it by using one of the unused bits. Simply extending the bitfield is
not binary compatible so this has to wait for Qt6.

Fixes: QTBUG-74652
Change-Id: I2539f5619ffe729459102fa5406fb6c0dd463ea1
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>